Backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
PyJWT>=2.8
werkzeug>=2.2
python-dotenv>=1.0
Authlib>=1.2
